+++ CHANGELOG.md
@@ -7,7 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `@rarimo/rarime` File Api
 - `@rarimo/rarime` new dev dependencies:
   - `@metamask/snaps-sdk`
   - `@metamask/snaps-utils`
@@ -15,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `@rarimo/rarime`:
   - Save credentials to ceramic instead of snap store
-  - Use File Api for MTPv2OnChain and Auth circuits
+  - Use MetaMask File API for `MTPv2OnChain` and Auth circuits
   - Return did in bigint representation alongside the default representation```
   - Updated dependencies:
     - `@metamask/snaps-jest` to "^4.0.0"