+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.0] - 2023-11-27
+## [1.0.1] - 2023-11-29
+### Fixed
+- `@rarimo/rarime` - `identity` creation
 
+## [1.0.0] - 2023-11-27
 ### Added
 - `@rarimo/rarime` new dev dependencies:
   - `@metamask/snaps-sdk`
@@ -117,7 +120,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Implemented `@rarimo/rarime-connector` and `@rarimo/rarime` packages
 
-[Unreleased]: https://github.com/rarimo/rarime/compare/0.8.0...HEAD
+[Unreleased]: https://github.com/rarimo/rarime/compare/1.0.1...HEAD
+[1.0.1]: https://github.com/rarimo/rarime/compare/1.0.0...1.0.1
+[1.0.0]: https://github.com/rarimo/rarime/compare/0.8.0...1.0.0
 [0.8.0]: https://github.com/rarimo/rarime/compare/0.7.2...0.8.0
 [0.7.2]: https://github.com/rarimo/rarime/compare/0.7.1...0.7.2
 [0.7.1]: https://github.com/rarimo/rarime/compare/0.7.0...0.7.1